// src/api/generated/models/listInternalTicketsApiV1TicketsInternalGetParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */
import type { SortOrder } from './sortOrder'
import type { TicketCategory } from './ticketCategory'
import type { TicketLifecycleFilter } from './ticketLifecycleFilter'
import type { TicketSortField } from './ticketSortField'
import type { TicketStatus } from './ticketStatus'
import type { TicketWorkflowState } from './ticketWorkflowState'

export type ListInternalTicketsApiV1TicketsInternalGetParams = {
  lifecycle?: TicketLifecycleFilter
  workflow_state?: TicketWorkflowState | null
  status?: TicketStatus | null
  category?: TicketCategory | null
  office_id?: string | null
  creator_user_id?: string | null
  primary_officer_id?: string | null
  current_assignee_id?: string | null
  sort_by?: TicketSortField
  order?: SortOrder
  created_from?: string | null
  created_to?: string | null
  updated_from?: string | null
  updated_to?: string | null
  /**
   * @minimum 1
   */
  page?: number
  /**
   * @minimum 1
   * @maximum 100
   */
  size?: number
  q?: string | null
}